fix schemas. closes #31

+++ shared/schema/build-schema.sh
@@ -14,17 +14,19 @@ SRC_XML_ARRAY=(
     "pat-amnd.xsl"
     "pat-rspn.xsl"
     "pat-etc.xsl"
+    "application-body.xsl"
     "foreign-language-body.xsl"
-    "common-templates/pat_common.xsl"
     "cpy-ntc-pt-e.xsl"
     "cpy-ntc-pt-e-rn.xsl"
     "cpy-ntc-pt-f.xsl"
+    "common-templates/pat_common.xsl"
     "v4xva_ntc-pt-e.xsl"
     "v4xva_ntc-pt-e-rn.xsl"
     "v4xva_ntc-pt-f.xsl"
     "common-templates/ntc-ninsyo.xsl"
     "common-templates/ntc-paragraph.xsl"
     "common-templates/dispatch-control-article.xsl"
+    "common-templates/unsupported-tags.xsl"
 )
 DSL="./src/schema-dsl.xsl"
 SAXON_JAR="/opt/saxon-he/saxon-he-10.8.jar"
@@ -75,5 +77,7 @@ for src_xml in "${SRC_XML_ARRAY[@]}"; do
          -xsl:"$DSL" \
          -s:"$XSL_ROOT/$src_xml" \
          -o:"$dst_json" --allowSyntaxExtensions:off 
-    jq . "$dst_json" > "${dst_json}.tmp" && mv "${dst_json}.tmp" "$dst_json"
+    if [ -f "$dst_json" ]; then
+        jq . "$dst_json" > "${dst_json}.tmp" && mv "${dst_json}.tmp" "$dst_json"
+    fi
 done
